move files & update relevant imports
